+++ modules/compiler/src/main/java/flex2/compiler/as3/binding/BindableFirstPassEvaluator.java
@@ -563,7 +563,7 @@ public class BindableFirstPassEvaluator extends GenerativeFirstPassEvaluator
 	private int getVisitedGetterSetterBindType(String name)
 	{
         return visitedProps == null || !visitedProps.containsKey(name) ? BINDABLE_NONE
-                : visitedProps.get(name).intValue();
+                : visitedProps.get(name);
 	}
 
 	/**
