+++ modules/compiler/src/main/java/flex2/compiler/mxml/builder/AnonymousObjectGraphBuilder.java
@@ -245,7 +245,7 @@ abstract class AnonymousObjectGraphBuilder extends AbstractBuilder
             }
             else
             {
-                int count = counts.get(localPart).intValue() + 1;
+                int count = counts.get(localPart) + 1;
                 counts.put(localPart, IntegerPool.getNumber(count));
             }
         }
@@ -272,7 +272,7 @@ abstract class AnonymousObjectGraphBuilder extends AbstractBuilder
             }
             else
             {
-                int num = counts.get(localPart).intValue() + 1;
+                int num = counts.get(localPart) + 1;
                 counts.put(localPart, IntegerPool.getNumber(num));
             }
         }
@@ -285,7 +285,7 @@ abstract class AnonymousObjectGraphBuilder extends AbstractBuilder
         Map<String, Array> arrays = new HashMap<String, Array>();
 
         for (String localPart : counts.keySet()) {
-            if (counts.get(localPart).intValue() > 1) {
+            if (counts.get(localPart) > 1) {
                 Array a = new Array(document, parent, parent.getXmlLineNumber(), typeTable.objectType);
                 a.setId(localPart, false);
                 a.setIsAnonymous(true);
