+++ modules/compiler/src/main/java/flex2/compiler/mxml/rep/MxmlDocument.java
@@ -825,7 +825,7 @@ public final class MxmlDocument
     {
         String typeName = NameFormatter.retrieveClassName(type.getName());
         Integer cell = anonIdCounts.get(typeName);
-        int i = cell == null ? 1 : cell.intValue();
+        int i = cell == null ? 1 : cell;
         anonIdCounts.put(typeName, i + 1);
         return i;
     }
